+++ example/lib/main.dart
@@ -1,182 +1,184 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 
-import 'package:open_settings_plus/android/open_settings_plus_android.dart';
-import 'package:open_settings_plus/ios/open_settings_plus_ios.dart';
+import 'package:open_settings_plus/open_settings_plus.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  static const settingsiOS = OpenSettingsPlusIOS();
-  static const settingsAndroid = OpenSettingsPlusAndroid();
-
   const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Plugin example app'),
+        body: NestedScrollView(
+          headerSliverBuilder: (context, innerBoxIsScrolled) => [
+            const SliverAppBar(
+              pinned: true,
+              backgroundColor: Colors.blue,
+              flexibleSpace: FlexibleSpaceBar(
+                title: Text("Open settings +"),
+              ),
+              centerTitle: true,
+              expandedHeight: 220,
+            ),
+          ],
+          body: switch (OpenSettingsPlus.shared) {
+            OpenSettingsPlusAndroid settings => _buildAndroidList(settings),
+            OpenSettingsPlusIOS settings => _buildIOSList(settings),
+            _ => const Center(
+                child: Text(
+                  "Unsupported platform.",
+                ),
+              ),
+          },
         ),
-        body: Builder(builder: (context) {
-          if (Platform.isAndroid) {
-            return _buildAndroidList();
-          } else if (Platform.isIOS) {
-            return _buildIOSList();
-          } else {
-            return const Center(
-              child: Text("Unsupported platform"),
-            );
-          }
-        }),
       ),
     );
   }
 
-  ListView _buildIOSList() {
+  ListView _buildIOSList(OpenSettingsPlusIOS settings) {
     return ListView(
       children: [
         ListTile(
-          onTap: settingsiOS.settings,
+          onTap: settings,
           title: const Text("Open settings"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.wifi,
+          onTap: settings.wifi,
           title: const Text("Open wi-fi"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.about,
+          onTap: settings.about,
           title: const Text("Open about"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.accessibility,
+          onTap: settings.accessibility,
           title: const Text("Open accessibility"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.accountSettings,
+          onTap: settings.accountSettings,
           title: const Text("Open account settings"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.autoLock,
+          onTap: settings.autoLock,
           title: const Text("Open auto lock"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.battery,
+          onTap: settings.battery,
           title: const Text("Open battery"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.bluetooth,
+          onTap: settings.bluetooth,
           title: const Text("Open bluetooth"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.dateAndTime,
+          onTap: settings.dateAndTime,
           title: const Text("Open date and time"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.faceIDAndPasscode,
+          onTap: settings.faceIDAndPasscode,
           title: const Text("Open face ID and passcode"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.cellular,
+          onTap: settings.cellular,
           title: const Text("Open cellular"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.dictionary,
+          onTap: settings.dictionary,
           title: const Text("Open dictionary"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.displayAndBrightness,
+          onTap: settings.displayAndBrightness,
           title: const Text("Open display and brightness"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.general,
+          onTap: settings.general,
           title: const Text("Open general"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.iCloud,
+          onTap: settings.iCloud,
           title: const Text("Open iCloud"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.music,
+          onTap: settings.music,
           title: const Text("Open music"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.keyboard,
+          onTap: settings.keyboard,
           title: const Text("Open keyboard"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.keyboards,
+          onTap: settings.keyboards,
           title: const Text("Open keyboards"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.languageAndRegion,
+          onTap: settings.languageAndRegion,
           title: const Text("Open language and region"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.locationServices,
+          onTap: settings.locationServices,
           title: const Text("Open location services"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.personalHotspot,
+          onTap: settings.personalHotspot,
           title: const Text("Open personal hotspot"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.phone,
+          onTap: settings.phone,
           title: const Text("Open phone"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.photosAndCamera,
+          onTap: settings.photosAndCamera,
           title: const Text("Open photos and camera"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.privacy,
+          onTap: settings.privacy,
           title: const Text("Open privacy"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.profilesAndDeviceManagement,
+          onTap: settings.profilesAndDeviceManagement,
           title: const Text("Open Profiles"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.storageAndBackup,
+          onTap: settings.storageAndBackup,
           title: const Text("Open storage and backup"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.siri,
+          onTap: settings.siri,
           title: const Text("Open siri"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsiOS.soundsAndHaptics,
+          onTap: settings.soundsAndHaptics,
           title: const Text("Open sounds and haptics"),
           trailing: const Icon(Icons.chevron_right),
         ),
@@ -184,76 +186,76 @@ class MyApp extends StatelessWidget {
     );
   }
 
-  ListView _buildAndroidList() {
+  ListView _buildAndroidList(OpenSettingsPlusAndroid settings) {
     return ListView(
       children: [
         ListTile(
-          onTap: settingsAndroid.wifi,
+          onTap: settings.wifi,
           title: const Text("Open wifi"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.accessibility,
+          onTap: settings.accessibility,
           title: const Text("Open accessibility"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.addAccount,
+          onTap: settings.addAccount,
           title: const Text("Open add account"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.airplaneMode,
+          onTap: settings.airplaneMode,
           title: const Text("Open airplane mode"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.apnSettings,
+          onTap: settings.apnSettings,
           title: const Text("Open apn settings"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.appNotification,
+          onTap: settings.appNotification,
           title: const Text("Open app notification"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.appSettings,
+          onTap: settings.appSettings,
           title: const Text("Open app settings"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.applicationDetails,
+          onTap: settings.applicationDetails,
           title: const Text("Open application details"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.applicationDevelopment,
+          onTap: settings.applicationDevelopment,
           title: const Text("Open application development"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.applicationNotification,
+          onTap: settings.applicationNotification,
           title: const Text("Open application notification"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.applicationSettings,
+          onTap: settings.applicationSettings,
           title: const Text("Open application settings"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.applicationWriteSettings,
+          onTap: settings.applicationWriteSettings,
           title: const Text("Open application write settings"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.batterySaver,
+          onTap: settings.batterySaver,
           title: const Text("Open battery saver"),
           trailing: const Icon(Icons.chevron_right),
         ),
         ListTile(
-          onTap: settingsAndroid.bluetooth,
+          onTap: settings.bluetooth,
           title: const Text("Open bluetooth"),
           trailing: const Icon(Icons.chevron_right),
         ),