+++ lib/core/open_settings_plus.dart
@@ -1,13 +1,16 @@
-import 'package:open_settings_plus/android/open_settings_plus_android.dart';
-import 'package:open_settings_plus/bridge/open_settings_plus_platform_interface.dart';
-import 'package:open_settings_plus/ios/open_settings_plus_ios.dart';
+import 'dart:io';
 
-/// {@template open_settings_plus_core}
+import 'package:open_settings_plus/bridge/open_settings_plus_platform_interface.dart';
+
+part 'open_settings_plus_android.dart';
+part 'open_settings_plus_ios.dart';
+
+/// {@template open_settings_plus}
 /// Singleton used to create a [OpenSettingsPlus] instance.
 /// You can get iOS or android instance of [OpenSettingsPlus] using this class.
 /// {@endtemplate}
-abstract class OpenSettingsPlus {
-  /// {@macro open_settings_plus_core}
+class OpenSettingsPlus {
+  /// {@macro open_settings_plus}
   const OpenSettingsPlus();
 
   /// Returns a singleton instance of [OpenSettingsPlus] for iOS.
@@ -15,6 +18,21 @@ abstract class OpenSettingsPlus {
 
   /// Returns a singleton instance of [OpenSettingsPlus] for android.
   const factory OpenSettingsPlus.android() = OpenSettingsPlusAndroid;
+
+  /// {@macro open_settings_plus_core}
+  static OpenSettingsPlus? build() {
+    if (Platform.isAndroid) {
+      return const OpenSettingsPlus.android();
+    } else if (Platform.isIOS) {
+      return const OpenSettingsPlus.iOS();
+    } else {
+      return null;
+    }
+  }
+
+  /// Singleton instance of [OpenSettingsPlus].
+  /// You can use this instance to open settings in iOS or android.
+  static final OpenSettingsPlus? shared = OpenSettingsPlus.build();
 
   /// Method used to pass custom message to native side.
   Future<bool> sendCustomMessage(String message) {