+++ lib/open_settings_plus.dart
@@ -1,3 +1,1 @@
-export 'android/open_settings_plus_android.dart';
-export 'core/open_settings_plus_core.dart';
-export 'ios/open_settings_plus_ios.dart';
+export 'core/open_settings_plus.dart';