+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+* Remove unnecessary comments.
+* Add a better documentation for the plugin.
+* Add nfc option to android.
+* Fix version of AGP8 and compilesdk version 34 (@prasant10050)
+* Add a better example for the plugin.
+* Better API by using singleton and switch case.
+
 ## 0.1.0
 * Add a stricter code analysis
 * Add NFC option to android.